Bug fix in device ptr
In this commit, we rearrange the pointer accesses such that
the device ptrs are set correctly and the allocation of
memory on device occurs correctly. However, a major drawback
of this fix is that it is not type safe now since in many
function signatures the 'const' qualifier had to be removed
to enable smooth integrations with cuda memory function signatures.